+++ inc/classes/ServiceProvider/class-common-subscribers.php
@@ -21,7 +21,6 @@ class Common_Subscribers extends AbstractServiceProvider {
 	 * @var array
 	 */
 	protected $provides = [
-		'heartbeat_subscriber',
 		'db_optimization_subscriber',
 		'capabilities_subscriber',
 		'webp_subscriber',
@@ -41,8 +40,6 @@ class Common_Subscribers extends AbstractServiceProvider {
 	public function register() {
 		$options = $this->getContainer()->get( 'options' );
 
-		$this->getContainer()->share( 'heartbeat_subscriber', 'WP_Rocket\Subscriber\Heartbeat_Subscriber' )
-			->withArgument( $options );
 		$this->getContainer()->share( 'db_optimization_subscriber', 'WP_Rocket\Subscriber\Admin\Database\Optimization_Subscriber' )
 			->withArgument( $this->getContainer()->get( 'db_optimization' ) )
 			->withArgument( $options );
